build: remove stray empty variable
Compilation can fail when libnfnetlink is not in a directory searched
by default. Reason is the empty KERNELDIR variable which makes for a
gcc command like:

	gcc -I. -I../include -I -Wall -I/usr/include/libnfnetlink-1.0.0+git28
	-Wall -c libnetfilter_queue.c

What one would expect is that gcc would search in the (non-existent)
directory "-Wall" and just continue as usual, since -Wall is specified
again. Instead, gcc versions before 4.6 attempt to search the
(similarly non-existent) directory "-I/usr/[...]" and thus miss.

Signed-off-by: Jan Engelhardt <jengelh@inai.de>