extra: pktbuff: pktb_expand_tail return 0 if there is no room in the tail
pktb_expand_tail returns 0 if there is no room for the mangling.
Note that we don't support dynamic reallocation, instead the
caller is responsible for allocating the extra room via pktb_alloc
according to the maximum amount of bytes it needs for the mangling.

Since pkt_buff layout is not exposed, we can change this in the
future if we prefer dynamic reallocation.

Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>